fix: Changed operator so that it does a server side apply rather than create for resources. (#470)

+++ crates/k8s-operator/src/services/pgadmin.rs
@@ -2,7 +2,7 @@ use super::deployment;
 use crate::error::Error;
 use k8s_openapi::api::apps::v1::Deployment;
 use k8s_openapi::api::core::v1::{ConfigMap, Secret, Service};
-use kube::api::{DeleteParams, PostParams};
+use kube::api::{DeleteParams, Patch, PatchParams, PostParams};
 use kube::{Api, Client};
 use serde_json::json;
 
@@ -12,17 +12,29 @@ const CONFIG_JSON: &str = include_str!("../../config/servers.json");
 // Large Language Model
 pub async fn deploy(
     client: Client,
-    password: &str,
-    keycloak_password: &str,
+    password: Option<String>,
+    keycloak_password: Option<String>,
     namespace: &str,
 ) -> Result<(), Error> {
+    // If we have the passwords then extract them.
+    let password = if let Some(password) = password {
+        password
+    } else {
+        "".to_string()
+    };
+    let keycloak_password = if let Some(keycloak_password) = keycloak_password {
+        keycloak_password
+    } else {
+        "".to_string()
+    };
     let passfile = format!("bionic-db-cluster-rw:5432:*:bionic_readonly:{}", password);
     let keycloak_passfile = format!(
         "keycloak-db-cluster-rw:5432:*:keycloak-db-owner:{}",
         keycloak_password
     );
-    // Put the envoy.yaml into a ConfigMap
-    let config_map = serde_json::from_value(serde_json::json!({
+
+    // Put the setup into a ConfigMap
+    let config_map = serde_json::json!({
         "apiVersion": "v1",
         "kind": "ConfigMap",
         "metadata": {
@@ -34,10 +46,18 @@ pub async fn deploy(
             "passfile": &passfile,
             "passfile_keycloak": &keycloak_passfile,
         }
-    }))?;
+    });
 
+    // If the config doesn't already exist then crate it.
     let api: Api<ConfigMap> = Api::namespaced(client.clone(), namespace);
-    api.create(&PostParams::default(), &config_map).await?;
+    if api.get(PGADMIN).await.is_err() {
+        api.patch(
+            PGADMIN,
+            &PatchParams::apply(crate::MANAGER),
+            &Patch::Apply(config_map),
+        )
+        .await?;
+    }
 
     deployment::deployment(
         client.clone(),
@@ -102,39 +122,55 @@ pub async fn deploy(
     )
     .await?;
 
-    let secret = serde_json::from_value(serde_json::json!({
-        "apiVersion": "v1",
-        "kind": "Secret",
-        "metadata": {
-            "name": PGADMIN,
-            "namespace": namespace
-        },
-        "stringData": {
-            "email": "pgadmin@pgadmin.com",
-            "password": super::database::rand_hex()
-        }
-    }))?;
+    pgadmin_secret(namespace, client).await?;
 
+    Ok(())
+}
+
+async fn pgadmin_secret(namespace: &str, client: Client) -> Result<(), Error> {
     let secret_api: Api<Secret> = Api::namespaced(client, namespace);
-    secret_api.create(&PostParams::default(), &secret).await?;
-
+    let secret = secret_api.get(PGADMIN).await;
+    if secret.is_err() {
+        let secret = serde_json::from_value(serde_json::json!({
+            "apiVersion": "v1",
+            "kind": "Secret",
+            "metadata": {
+                "name": PGADMIN,
+                "namespace": namespace
+            },
+            "stringData": {
+                "email": "pgadmin@pgadmin.com",
+                "password": super::database::rand_hex()
+            }
+        }))?;
+        secret_api.create(&PostParams::default(), &secret).await?;
+    }
     Ok(())
 }
 
 pub async fn delete(client: Client, namespace: &str) -> Result<(), Error> {
     // Remove deployments
     let api: Api<Deployment> = Api::namespaced(client.clone(), namespace);
-    api.delete(PGADMIN, &DeleteParams::default()).await?;
+
+    if api.get(PGADMIN).await.is_ok() {
+        api.delete(PGADMIN, &DeleteParams::default()).await?;
+    }
 
     let api: Api<ConfigMap> = Api::namespaced(client.clone(), namespace);
-    api.delete(PGADMIN, &DeleteParams::default()).await?;
+    if api.get(PGADMIN).await.is_ok() {
+        api.delete(PGADMIN, &DeleteParams::default()).await?;
+    }
 
     // Remove services
     let api: Api<Service> = Api::namespaced(client.clone(), namespace);
-    api.delete(PGADMIN, &DeleteParams::default()).await?;
+    if api.get(PGADMIN).await.is_ok() {
+        api.delete(PGADMIN, &DeleteParams::default()).await?;
+    }
 
     let secret_api: Api<Secret> = Api::namespaced(client, namespace);
-    secret_api.delete(PGADMIN, &DeleteParams::default()).await?;
+    if api.get(PGADMIN).await.is_ok() {
+        secret_api.delete(PGADMIN, &DeleteParams::default()).await?;
+    }
 
     Ok(())
 }
